Move llm and rag related code to core/
