+++ src/core/llm/exception.ts
