+++ src/core/llm/manager.ts
