+++ src/core/llm/ollama.ts
