+++ src/core/llm/openai.ts
