+++ src/core/llm/openaiCompatibleProvider.ts
